Fix Mech's Guns & Upper Bodies not rotating correctly
trial and error baby

guns and upper bodies are dumb and stupid when mirroring the player on
the x axis xor the z axis. they are NOT dumb and stupid when both
applied.
to make the gun not dumb and stupid, we need to tell it that it is
facing 180 degrees more than it thinks it is. idk why this works but it
does.
to make the upper body not dumb and stupid, we need to tell it to be
mirrored, that is negated from 360 degrees. idk why this works either.

moral of the story, spend a dozen hours looking for something and you
might find it.